Move some of the more involved instruction implementations to their own methods
This was already the case for PopCall - it's a biggun so it lives in its
own method for readability. GetAttr and SetAttr have been moved to their
own respective methods as well. StoreReturn was in its own method, but
it was pretty short, so that's been factored out and the body of
store_return is back to the main loop.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>