Add compile::thunk mod and some new jumping instructions
WIP: The road to turing completeness

Thunks are a basic block abstraction that are constructed with
instructions at the most basic level. The ir_to_instr transator has also
been updated to use them, since we're only making non-jumping code so
far.

The additional jump instructions haven't been implemented in the VM, but
it's a simple collection of:

* PopCmp - pop and compare truthiness, setting some internal VM flag
* Jump(x) - jump to address X unconditionally
* JumpTrue(x) - jump to address X if the last compare was true
* JumpFalse(x) - jump to address X if the last compare was false

I'm hoping to implement greater than, less than, etc with native
functions or builtin functions at the worst.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>