Squash minor warning that was popping up
There's lots of warnings for compile right now, but this one was a
little more urgent because it was due to deprecated behavior. Calling
functions is a little fragile right now because of how and where
functions are stored alongside `self` being borrowed mutably in this
situation. I may look at an unsafe solution, but function call clones
are relatively cheap so it's not a huge deal to clone it explicitly
here.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>