Add vm::state module, add Type object to make_types, and more
* Builtin functions may require calling function references - so they
now accept a reference to the VM's state, instead of the storage. This
way, functions may be called from other functions in a more-or-less
direct fashion.
* Type objects in compile::object contain stuff like the __make__
construction method and __call__ for handling when an object is called.
* MakeTypes is being invoked by the compiler now

I want to refactor and re-organize the code because I'm shoving a lot of
stuff into the same file and it's starting to get crowded. That will
probably be next.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>