Add compile-time builtins, restore IR atom translation (sort of)
For a while I was trying to define builtins at the very end of the
compile-time chain, but after struggling and writing some seriously ugly
code (see: src/compile/object.rs, it's still in there) I think the best
approach was to just treat builtins as if they were sort of syntax
items, but still with special gloves to make sure that we aren't
introducing weird stuff into the main codebase. So now builtins are
collected in a similar, separate step from AST collection. This is also
where builtin types and functions are defined, so as far as organization
goes, it kind of doesn't make sense. When re-writing this part, the
problem I kept running in to was trying to avoid repeating defining
builtin structure definitions and builtin structure members; I may have
to take a few ideas from the now defunct compile/object.rs source.

The only other notable change is that translate/ir.rs has its
translate_atom method uncommented, so... yippee.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>