Make IR context in ir translator a mut reference
The context (IrCtx object) in the IR to artifact translator was
previously owned by the IrToInst translator. However, this is not
conducive for recursive translation in the functions and types.
Functions need to be able to translate their IR in the same context, so
using references makes that a whole lot easier.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>