Update object system and add constant translation skeleton
* Object system with StringObject being keys simply won't work. This is
because they're backed by BaseObjects, which require having a ValueRef,
aka a value in the constants or heap. However, the StringObject was
owned by the map as its keys - thus, it couldn't be addressed using one
of those refs. So instead, base objects use base Rust strings.
* Constant translation has its type system in place, even if it's a
little clunky for bootstrapping. Constant objects will be created and
then hopefully a VM artifact will be able to be created.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>