Update register_type_with to use get_builtin_binding instead of get_binding
register_type_with in constant translation is used specifically for
registering builtin types, and looks up builtin bindings using its
string name. In the case that something in the main function body
defines a type with the same name as a builtin, this would have used
that instead of the builtin type name. This change fixes that.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>